pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dlab"
version = "2.0.0"
requires-python = ">=3.10"
dependencies = [
  "PyQt5>=5.15.0",
  "pyyaml>=6.0",
  "numpy",
  "matplotlib",
  "prometheus-client",
  "pyserial",
  "prysm",
  "pylablib",
  "ThorlabsPM100",
  "thorlabs-apt",
  "pyvisa",
  "scipy"
]

[project.scripts]
dlab = "dlab.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["dlab*"]
